+++ pubhubs/src/misc/mod.rs
@@ -4,6 +4,7 @@ pub mod drop_ext;
 pub mod error;
 pub mod fmt_ext;
 pub mod jwt;
+pub mod net_ext;
 pub mod serde_ext;
 pub mod task;
 pub mod time_ext;